--- app/src/main/java/cord/eoeo/momentwo/data/MomentwoPagingSource.kt
@@ -1,29 +1,0 @@
-package cord.eoeo.momentwo.data
-
-import androidx.paging.PagingSource
-import androidx.paging.PagingState
-import retrofit2.HttpException
-import java.io.IOException
-
-abstract class MomentwoPagingSource<T : Any> : PagingSource<Int, T>() {
-    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, T> {
-        val page = params.key ?: 0
-
-        return try {
-            loadPage(page)
-        } catch (exception: IOException) {
-            return LoadResult.Error(exception)
-        } catch (exception: HttpException) {
-            return LoadResult.Error(exception)
-        }
-    }
-
-    override fun getRefreshKey(state: PagingState<Int, T>): Int? {
-        return state.anchorPosition?.let { anchorPosition ->
-            state.closestPageToPosition(anchorPosition)?.prevKey?.plus(1)
-                ?: state.closestPageToPosition(anchorPosition)?.nextKey?.minus(1)
-        }
-    }
-
-    protected abstract suspend fun loadPage(page: Int): LoadResult<Int, T>
-}
